Create v0.0
This is a prototype version of v1.0 that I made. This version while still usable, is built upon the idea that the user will only define a model with three target variables. The bits of hard coded values in the program make it ill-suited for anything other than a very specific task, which is what prompted me to develop a program that accepts up to five target variables in any order.